use 'ChooseRandomSingleLegModeForSubtour" which is in this package and not the one in the core.
This change shouldn't have a any big effect since the only thing which
is  done differently is that the route is always set to null. This also
means, we can get rid of our 'own' subtour mode choice classes in this
repository. The only thing which matters is that we keep the
SubtourModeChoiceRepairReRoute strategy which only repairs routes after
doing mode choice instead of re-routing all routes.